bug fix - test if event exists before destroy

+++ db/migrate/20171013072017_add_2017_local_party.rb
@@ -30,6 +30,6 @@ class Add2017LocalParty < ActiveRecord::Migration
 
   def down
     election = Election.includes(:election_translations).where(election_translations: {name: '2017 Local Election - Party List'}).first
-    election.destroy
+    election.destroy if election.present?
   end
 end
